Added prohibition on editing/deleting iFeeds when not being the owner and/or having the iFeed role. Now with check on the server also. Removed LR-specific exception signature on utility method.

+++ trunk/core-bc/modules/portlet/src/main/java/se/vgregion/ifeed/controller/RemoveIFeedController.java
@@ -1,6 +1,7 @@
 package se.vgregion.ifeed.controller;
 
 import javax.portlet.ActionRequest;
+import javax.portlet.PortletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.portlet.bind.annotation.ActionMapping;
 
+import se.vgregion.ifeed.el.AccessGuard;
 import se.vgregion.ifeed.service.ifeed.IFeedService;
 import se.vgregion.ifeed.types.IFeed;
 
@@ -15,8 +17,10 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.model.ResourceConstants;
+import com.liferay.portal.model.User;
 import com.liferay.portal.service.ResourceLocalService;
 import com.liferay.portal.theme.ThemeDisplay;
+import com.liferay.portal.util.PortalUtil;
 
 @Controller
 @RequestMapping("VIEW")
@@ -31,7 +35,13 @@ public class RemoveIFeedController {
 	}
 
 	@ActionMapping(params = "action=removeIFeed")
-	public void removeBook(@RequestParam final Long feedId, final ActionRequest request) {
+	public void removeBook(@RequestParam final Long feedId, final ActionRequest request, final PortletRequest pr)
+	        throws PortalException, SystemException {
+		User user = PortalUtil.getUser(request);
+		IFeed feed = iFeedService.getIFeed(feedId);
+		if (!AccessGuard.mayEditFeed(user, feed)) {
+			throw new RuntimeException();
+		}
 		iFeedService.removeIFeed(feedId);
 		try {
 			ThemeDisplay themeDisplay = (ThemeDisplay) request.getAttribute(WebKeys.THEME_DISPLAY);
